refactor(presets): unify graphql operation resolution
